feat: Go実装からwriteSSEを削除
